+++ kytos/core/api_server.py
@@ -11,38 +11,30 @@ from flask_socketio import SocketIO
 class APIServer:
     """Api server used to provide Kytos Controller routes."""
 
-    def __init__(self, app_name, debug=False):
+    def __init__(self, app_name, listen='0.0.0.0', port=8181):
         """Constructor of APIServer.
 
         This method will instantiate a server with SocketIO+Flask.
 
         Parameters:
             app_name(string): String representing a App Name
+            listen (string): host name used by api server instance
+            port (int): Port number used by api server instance
         """
         dirname = os.path.dirname(os.path.abspath(__file__))
         self.flask_dir = os.path.join(dirname, '../web-ui')
         self.log = logging.getLogger('werkzeug')
-        self.log.setLevel(logging.WARNING)
-        self.set_debug(debug)
+
+        self.listen = listen
+        self.port = port
 
         self.app = Flask(app_name, root_path=self.flask_dir)
         self.server = SocketIO(self.app, async_mode='threading')
 
-    def set_debug(self, debug):
-        """Method used to set debug mode.
-
-        Args:
-            debug(bool): Boolean value to turn on/off debug mode.
-        """
-        if debug:
-            self.log.setLevel(logging.DEBUG)
-        else:
-            self.log.setLevel(logging.WARNING)
-
-    def run(self, *args, **kwargs):
+    def run(self):
         """Method used to run the APIServer."""
         try:
-            self.server.run(self.app, *args, **kwargs)
+            self.server.run(self.app, self.listen, self.port)
         except OSError:
             pass
 
@@ -101,11 +93,11 @@ class APIServer:
         """Display json with kytos status using the route '/kytos/status/'."""
         return '{"response": "running"}', 201
 
-    @staticmethod
-    def stop_api_server():
+    def stop_api_server(self):
         """Method used to send a shutdown request to stop Api Server."""
         try:
-            urlopen('http://127.0.0.1:8181/kytos/shutdown/')
+            url = 'http://{}:{}/kytos/shutdown'.format('127.0.0.1', self.port)
+            urlopen(url)
         except URLError:
             pass
 
@@ -115,7 +107,8 @@ class APIServer:
         This method must be called by kytos using the method
         stop_api_server, otherwise this request will be ignored.
         """
-        allowed_host = ['127.0.0.1:8181', 'localhost:8181']
+        allowed_host = ['127.0.0.1:'+str(self.port),
+                        'localhost:'+str(self.port)]
         if request.host not in allowed_host:
             return "", 403